workflow: move committing workflows to schedule
This moves all workflows that commit to a schedule, so that it _doesn't_
push into peoples PRs as that enlarges them and is just non-obvious.

They run throughout the week.

Also change the "user" we use for this, so that DCO believes it a real
email address (DCO is pretty stupid here). Update the authors target to
filter out this bot as well.

Signed-off-by: Miek Gieben <miek@miek.nl>