plugin/cache: add refresh_stale parmeter
The `refresh_stale` parameter can be set to `sync` in order
to ensure that TTL is respected when the upstream is available.
This does not negate cache benefits,q because refresh will
not occur until the TTL expires. If the upstream is not
responsive, then `sync` will delay the response until the
upstream request has failed. This behavior is useful for
cases where the extra request time introduced by an
unhealthy upstream would be less harmful than it would be
to receive an expired entry which is no longer valid.

* `refresh_stale` specifies a refresh mode for `serve_stale`. The default is `async`.
  * `async` mode will return an expired entry and attempt to refresh it asynchronously
  * `sync` mode will synchronously refresh an expired entry, and will serve an expired entry only after refresh has failed or exceeded the timeout.

See #4097

Signed-off-by: Zac Medico <zmedico@gmail.com>