Fix timestamp issue for doc generating
This PR will fix the issue of doc is not automated rendered
through GitHub Actions anymore.

It looks the issue is cause by the fact that in `fixup_file_mtime.sh`
is only fixing files on source (.md) side, not on target (man/*.[1|5|7])
side. As a result Makefile will skip the rendering of doc as
it assume everything will be update to date.

This should fix the issue we were facing.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>