plugin/log: Add option to log slow dns responses
This change adds a flag to the log plugin to log slow dns responses.
The duration of what is defined as slow is configurable, default is
100ms.
Slow dns responses are important to keep track of as dns clients
may have a low timeout and a late response may therefore be considered
as no response whatsoever.

Signed-off-by: Johannes Scheerer <johannes.scheerer@sap.com>