[plugin/clouddns] Add additional doc in explaining credentials file field is optional
This PR adds additional doc in explaining credentials file field is optional,
in case user might be concerned to save an filename location in Corefile.

Technically since Corefile does not save plaintext secret for clouddns
(unlikely route53/azure), this is not too big of a concern, as far as I can see.
Still it is worth to pointing out in documentation.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>