plugin/k8s_external: Fix rcode for headless services
The TestExternal test for the kubernetes plugin had an erroneous return
statement that prevented the test from verifying any but the first test
case.  This commit removes that bogus return statement.

The test cases also had several typos, which this commit corrects.

Fixing TestExternal revealed that External was returning the wrong rcode in
some cases.  Instead of returning NOERROR, External was returning NXDOMAIN
for headless services.  This commit corrects that error.

* plugin/kubernetes/external.go (External): Fix rcode for headless
services.
* plugin/kubernetes/external_test.go (TestExternal): Verify all test cases.

Signed-off-by: Miciah Masters <miciah.masters@gmail.com>