plugin/etcd: fix NS records when PathPrefix contains multiple path components
The `plugin/stcd/msg.Domain()` function currently operates on the
assumption that the prefix always contains a single path component:
`/foo` or `foo`, not `foo/bar`. However, nothing prohibits the etcd
plugin from being configured with a PathPrefix like `/skydns/foo`.

When the `etcd` plugin is configured this way, NS queries expose all of
the path components after the first one. So a prefix of `/skydns/foo`
and a zone of `bar.local` results in an NS record at
`/skydns/foo/local/bar/dns/ns` and returns `ns.dns.bar.local.foo.` as
the nameserver, where the correct answer would be `ns.dns.bar.local.`.

To handle these cases, we add a new `RName` property to the `Service`
struct, which allows the plugin to override the name that would be
generated by `msg.Domain()`. The method `.Domain()` on that struct
returns the `RName` property if it's non-empty, and otherwise calls
`msg.Domain()` to invoke the previous default behavior.

At present, the etcd plugin is the only plugin that has been updated to
override RName. Other plugins may need this fixed as well; a new
DomainWithPrefix public helper function is exported from `msg` to fit
this use case.

The existing `TestOtherLookup` test case has been modified to verify the
new etcd plugin behavior.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>