[DoH] Make protocol parsing more flexible
This change avoid the hard coding of HTTPS as the protocol, which will e.g. make
testing in external libraries easier and it will allow for additional
flexibility in whether HTTP or HTTPS should be used for the call.

The change is also updating the test to be a table based test with more
scenarios and also utilizing Go sub tests for nicer test output.

Signed-off-by: Sebastian Dahlgren <sebdah@fb.com>