plugin/cache: add a new keepttl option
Whenever CoreDNS is used as an authoritative server it could be useful
to serve a consistent TTL (not decreased based on time) to clients and
still use caching to spare request to the backend.

This commit adds a new option `keepttl` to the cache plugin to do just
that. This new option doesn't update any caching logic but just override the
cache plugin answer.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@proton.ch>