enabled query name and type as part of the metrics optionally
As part of this change, we are introducing a new configuration for the
`forward` plugin named `qtype_for_metrics` that can be used to configure
the query types that needs to be generated into metrics which contains
the additional information of `name` and `type`.

The new metrics will be `request_duration_seconds_with_name`
and will contain `"proxy_name", "to", "rcode", "name", "type"` as the
labels/tags in the metrics.

Signed-off-by: Harsha Narayana <harshnar@cisco.com>