fix: `Dockerfile` - Remove `setcap` stage
When the image was migrated to run as a non-root user some container engines had not yet relaxed `cap_net_bind_service` by default to permit non-root container users to bind to ports below 1024. Today it is common to find this capability for containers configured to `0`.

Using `setcap` with `e` enforces the capability requirement via a kernel check before the binary itself is executed, failing to run if the capability has been forbidden (_even when it is not required if the container were configured to run as the container user as root_). This type of workaround is considered ["capability-dumb"](https://man7.org/linux/man-pages/man7/capabilities.7.html), instead of properly handling it by the application itself at runtime to raise the capability from the permitted set into the effective set.

As such, those who are more security conscious and drop all capabilities by default must explicitly add capabilities a program enforces, even when they should not be required for the runtime configuration to succeed.

The correct approach would have been to use `setcap` with only `p` (permitted set), with application code raising the capability as necessary (_in this case the CoreDNS process is run unprivileged and binding to a port below the associated sysctl `net.ipv4.ip_unprivileged_port_start`_).

---

The remainder of the removed stage was intended to improve security, however this is questionable given:
- As has been previously communicated to maintainers, the certs copied have been demonstrated to be unaffected by the `upgrade`.
- The delta between an upgrade to packages being available since the last published revision of the base image tag is smaller than the frequency of CoreDNS image tags being updated, notably weakening the justification.
- Furthermore, given the multi-stage `Dockerfile` relying on `COPY` for the CoreDNS binary instead of building it within the container, and that none of the upgraded content of the images beyond the certificates were being carried over into the final runtime stage, there is nothing to gain from the upgrade step.

As such, it is better to drop the stage entirely in favor of keeping the `Dockerfile` simple and focused.

---

The default platform for `FROM` should be the target platform, so that has been dropped as well.

Signed-off-by: Brennan Kinney <5098581+polarathene@users.noreply.github.com>